Implements Zoom Functionality on Top Left Button
Adds zoom functionality by mapping the top left button to Ctrl+Scroll.

This allows the user to perform smooth zoom operations in applications by holding the top left button and moving the trackball up or down.

Reduces the number of taps required for Home function on the mid left button from 3 to 2. Similarly, reduces the number of taps required for End function on the mid left button from 4 to 3.